Salim

+++ python_async_function/2-measure_runtime.py
@@ -7,7 +7,7 @@ import asyncio
 wait_n = __import__('1-concurrent_coroutines').wait_n
 
 
-def measure_time(n: int, max_delay: int) -> int:
+def measure_time(n: int, max_delay: int) -> float:
     """Measures the average execution time for wait_n"""
 
     start = time.time()
